src/main/java/ru/job4j/tracker/DeleteAction.java
package ru.job4j.tracker;

public class DeleteAction implements UserAction {
    private final Output out;

    public DeleteAction(Output out) {
        this.out = out;
    }

    @Override
    public String name() {
        return "Delete";
    }

    @Override
    public boolean execute(Input input, Store store) {
        out.println("=== Delete item ===");
        int id = input.askInt("Enter id ");
        if (store.delete(id)) {
            out.println("Item delete is done");
        } else {
            out.println("Item delete error");
        }
        return true;
    }
}
